Tela de consultar tds as denuncias
