--- src/main/java/view/TelaConsultarDenuncias1.java
@@ -1,339 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package view;
-
-import controller.DenunciaService;
-import model.Denuncia;
-
-import javax.swing.*;
-import javax.swing.table.DefaultTableModel;
-import java.util.ArrayList;
-import java.util.List;
-
-import model.Usuario;
-
-/**
- * @author User
- */
-public class TelaConsultarDenuncias1 extends javax.swing.JFrame {
-
-    private Denuncia denuncia;
-
-    /**
-     * Creates new form TelaConsultarDenunciasU
-     */
-    public TelaConsultarDenuncias1() {
-        initComponents();
-        this.denuncia = new Denuncia();
-        this.carregarTabela();
-        saveButton.setVisible(false);
-    }
-
-    private javax.swing.JTextField campoId;
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(TelaConsultarDenuncias1.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(TelaConsultarDenuncias1.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(TelaConsultarDenuncias1.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(TelaConsultarDenuncias1.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new TelaConsultarDenuncias1().setVisible(true);
-            }
-        });
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jScrollPane1 = new javax.swing.JScrollPane();
-        jTableDenuncias = new javax.swing.JTable();
-        jScrollPane2 = new javax.swing.JScrollPane();
-        campoDesc = new javax.swing.JTextArea();
-        campoLoc = new javax.swing.JTextField();
-        campoId = new javax.swing.JTextField();
-        editButton = new javax.swing.JButton();
-        eraseButton = new javax.swing.JButton();
-        backButton = new javax.swing.JButton();
-        jLabel1 = new javax.swing.JLabel();
-        saveButton = new javax.swing.JButton();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.DISPOSE_ON_CLOSE);
-
-        jTableDenuncias.setModel(new javax.swing.table.DefaultTableModel(
-
-            new Object [][] {
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null}
-            },
-            new String [] {
-                "ID", "Localização Bairro/Rua/Comp", "Desc."
-            }
-
-        ) {
-            Class[] types = new Class[]{
-                    java.lang.Integer.class, java.lang.String.class, java.lang.String.class
-            };
-            boolean[] canEdit = new boolean[]{
-                    false, false, false
-            };
-
-            public Class getColumnClass(int columnIndex) {
-                return types[columnIndex];
-            }
-
-            public boolean isCellEditable(int rowIndex, int columnIndex) {
-                return canEdit[columnIndex];
-            }
-        });
-        jTableDenuncias.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                jTableDenunciasMouseClicked(evt);
-            }
-        });
-        jScrollPane1.setViewportView(jTableDenuncias);
-        if (jTableDenuncias.getColumnModel().getColumnCount() > 0) {
-            jTableDenuncias.getColumnModel().getColumn(0).setResizable(false);
-            jTableDenuncias.getColumnModel().getColumn(1).setResizable(false);
-            jTableDenuncias.getColumnModel().getColumn(2).setResizable(false);
-        }
-
-        campoDesc.setColumns(20);
-        campoDesc.setLineWrap(true);
-        campoDesc.setRows(5);
-        jScrollPane2.setViewportView(campoDesc);
-
-        editButton.setText("Editar denúncia");
-        editButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                editButtonActionPerformed(evt);
-            }
-        });
-
-        eraseButton.setText("Apagar denúncia");
-        eraseButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                eraseButtonActionPerformed(evt);
-            }
-        });
-
-        backButton.setText("Voltar");
-        backButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                backButtonActionPerformed(evt);
-            }
-        });
-
-        jLabel1.setFont(new java.awt.Font("Segoe UI", 1, 14)); // NOI18N
-        jLabel1.setText("Dados:");
-
-        saveButton.setText("Salvar");
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                        .addGroup(layout.createSequentialGroup()
-                                .addGap(30, 30, 30)
-                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                                        .addGroup(layout.createSequentialGroup()
-                                                .addComponent(backButton, javax.swing.GroupLayout.PREFERRED_SIZE, 95, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                                .addComponent(jScrollPane2, javax.swing.GroupLayout.PREFERRED_SIZE, 353, javax.swing.GroupLayout.PREFERRED_SIZE))
-                                        .addGroup(layout.createSequentialGroup()
-                                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                                        .addComponent(editButton, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                                        .addComponent(eraseButton, javax.swing.GroupLayout.DEFAULT_SIZE, 186, Short.MAX_VALUE))
-                                                .addGap(18, 18, 18)
-                                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 805, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                        .addGroup(layout.createSequentialGroup()
-                                                                .addGap(6, 6, 6)
-                                                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                                                        .addComponent(campoLoc, javax.swing.GroupLayout.PREFERRED_SIZE, 414, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                                        .addComponent(jLabel1)
-                                                                        .addComponent(campoId, javax.swing.GroupLayout.PREFERRED_SIZE, 414, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                                                        .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                                                                .addComponent(saveButton, javax.swing.GroupLayout.PREFERRED_SIZE, 96, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                                .addGap(385, 385, 385)))))
-                                .addGap(29, 29, 29))
-        );
-        layout.setVerticalGroup(
-                layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                        .addGroup(layout.createSequentialGroup()
-                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                        .addGroup(layout.createSequentialGroup()
-                                                .addGap(26, 26, 26)
-                                                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 404, javax.swing.GroupLayout.PREFERRED_SIZE))
-                                        .addGroup(layout.createSequentialGroup()
-                                                .addGap(63, 63, 63)
-                                                .addComponent(editButton, javax.swing.GroupLayout.PREFERRED_SIZE, 58, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                .addGap(18, 18, 18)
-                                                .addComponent(eraseButton, javax.swing.GroupLayout.PREFERRED_SIZE, 58, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                                .addGap(19, 19, 19)
-                                .addComponent(jLabel1)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                        .addGroup(layout.createSequentialGroup()
-                                                .addGap(108, 108, 108)
-                                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                                                        .addComponent(backButton)
-                                                        .addComponent(saveButton)))
-                                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                                .addComponent(jScrollPane2)
-                                                .addGroup(layout.createSequentialGroup()
-                                                        .addComponent(campoLoc, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                        .addGap(18, 18, 18)
-                                                        .addComponent(campoId, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                        .addGap(0, 0, Short.MAX_VALUE))))
-                                .addContainerGap(23, Short.MAX_VALUE))
-        );
-
-        pack();
-        setLocationRelativeTo(null);
-    }// </editor-fold>//GEN-END:initComponents
-
-    public void carregarTabela() {
-
-        DefaultTableModel model = (DefaultTableModel) this.jTableDenuncias.getModel();
-        model.setNumRows(0);
-
-        DenunciaService denunciaService = new DenunciaService();
-        List<Denuncia> listaMinhasDenuncias = denunciaService.listarDenuncia();
-
-        for (Denuncia d : listaMinhasDenuncias) {
-            model.addRow(new Object[]{
-                    d.getId(),
-                    d.getLocalizacaoC(),
-                    d.getDescricao()
-            });
-        }
-
-    }
-
-    private void disableCampos() {
-        campoLoc.setEnabled(false);
-        campoDesc.setEnabled(false);
-    }
-
-    private void enableCampos() {
-        campoLoc.setEnabled(true);
-        campoDesc.setEnabled(true);
-    }
-    
-    private void eraseButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_eraseButtonActionPerformed
-        // TODO add your handling code here:
-        int option = JOptionPane.showConfirmDialog(this, "Deseja realmente apagar a denúncia?", "Confirmação", JOptionPane.YES_NO_OPTION);
-        int id = 0;
-        DenunciaService denunciaService = new DenunciaService();
-        if (option == JOptionPane.YES_OPTION) {
-            id = Integer.parseInt(this.campoId.getText());
-            denunciaService.apagarDenuncia(id);
-            this.carregarTabela();
-        }
-    }//GEN-LAST:event_eraseButtonActionPerformed
-
-    private void backButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_backButtonActionPerformed
-        // TODO add your handling code here:
-        this.dispose();
-
-    }//GEN-LAST:event_backButtonActionPerformed
-
-    private void editButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_editButtonActionPerformed
-        // TODO add your handling code here:
-        saveButton.setVisible(true);
-        enableCampos();
-
-
-    }//GEN-LAST:event_editButtonActionPerformed
-
-    private void saveButtonActionPerformed(java.awt.event.ActionEvent evt) {                                           
-        // TODO add your handling code here:
-        saveButton.setVisible(true);
-
-        //com o botão editar ele autera a denúncia no banco de dados (leo não ciou uma ação para o botão salvar)
-        int option = JOptionPane.showConfirmDialog(this, "Deseja realmente editar a denúncia?", "Confirmação", JOptionPane.YES_NO_OPTION);
-        int id;
-        String loc;
-        String desc;
-        DenunciaService denunciaService = new DenunciaService();
-
-        if (option == JOptionPane.YES_OPTION) {
-            loc = this.campoLoc.getText();
-            desc = this.campoDesc.getText();
-            id = Integer.parseInt(this.campoId.getText());
-            Denuncia denuncia = new Denuncia(id, desc, loc);
-
-            denunciaService.editarDenuncia(denuncia);
-            this.carregarTabela();
-            disableCampos();
-        }
-
-    }          
-    
-    // Variables declaration - do not modify                     
-    private javax.swing.JButton backButton;
-    private javax.swing.JTextArea campoDesc;
-    private javax.swing.JTextField campoLoc;
-    private javax.swing.JButton editButton;
-    private javax.swing.JButton eraseButton;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JScrollPane jScrollPane2;
-    private javax.swing.JTable jTableDenuncias;
-    private javax.swing.JButton saveButton;
-
-    private void jTableDenunciasMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_jTableDenunciasMouseClicked
-        // TODO add your handling code here:
-        if (this.jTableDenuncias.getSelectedRow() != -1) {
-            int id = (int) this.jTableDenuncias.getValueAt(this.jTableDenuncias.getSelectedRow(), 0);
-            String loc = this.jTableDenuncias.getValueAt(this.jTableDenuncias.getSelectedRow(), 1).toString();
-            String desc = this.jTableDenuncias.getValueAt(this.jTableDenuncias.getSelectedRow(), 2).toString();
-
-
-            this.campoLoc.setText(loc);
-            this.campoDesc.setText(desc);
-            this.campoId.setText(String.valueOf(id));
-
-
-        }
-    }//GEN-LAST:event_jTableDenunciasMouseClicked
-
-    // End of variables declaration                   
-}
